Moved Cluster auth test to IT
